Prevent bufwinnr() from showing an error message
When 'debug' is set, bufwinnr() displays an error message
saying the thesaurus buffer is missing if the thesaurus window
is closed.
According to the documentation, this function returns -1 when
a buffer's window can't be found.
This value is checked for and acted upon properly,
so there is no need for that error message, especially
when it bothers people (#25)